feat(s9): adopt the value objects, with converters keeping the stored and wire shapes fixed
Wires the S9 kernel into the entities, DTOs, mappers and services, and adds the
two converter classes that make it a compile-time-only change.

DomainTypeConverters is the load-bearing piece: each pair writes back exactly the
BSON type the documents already hold — Money -> double, Quantity/Rating -> int,
the string-backed ids and value objects -> String, CategoryId -> ObjectId.
Registering a writing converter is also what keeps Money a "simple" type; without
one Spring Data would store it as a nested {amount, currency} document. Reading
accepts any Number, because the seed data holds integral prices as int32
("originalPrice": 195) next to doubles.

ObjectIdBackedIdConverters covers what a global converter cannot express:
ProductId is stored as a String in orders.items[] and carts.items[] but as an
ObjectId in reviews.productId — one Java type, two BSON forms. MongoCustomConversions
registers per type; @ValueConverter applies per property, so Review's three id
fields opt out of the global pairs and keep their ObjectId form, which the rating
distribution aggregation joins on.

Fields adopted: Money on Order.totalAmount, OrderItem.price, CartItem.price,
Product.price/originalPrice and Transaction.amount; CategoryId on
Product.categoryIds; ProductCondition on Product.condition and CartItem.condition;
Role on User.role; ProductId/OrderId/UserId on Review's three ObjectId fields.
Money defaults to ZERO on every adopted field so a document missing it still reads
as 0.00, exactly as the primitive double did.

The order-total arithmetic is preserved exactly, not approximately: every operand
was already effectively 2dp, and the characterization suite still asserts 36.60,
43.20, 47.40, 18.45 and 12.50 — the same numbers, now typed. CustomerOrderService
no longer round-trips through doubleValue().

PersonName closes the fullName split. User.setName writes the three stored fields
together, AuthService.register derives the name from firstName/lastName instead of
trusting the client's fullName, and UserService.updateProfile stops recomputing it
by hand. UserRegisterRequestDTO.fullName is now @Deprecated and no longer
@NotBlank: still accepted so both frontends keep working, but ignored. OAuth keeps
its previous behaviour for a one-word Google display name via PersonName.tryParse.

Wire format, verified by literal-JSON assertions rather than by the DTO-equals-entity
tests (which would have passed even if Money had serialized as an object):
  price 1599.0, originalPrice 120.5, totalAmount 3198.0 — still JSON doubles
  condition "LIKE NEW" — the spaced token both frontends key on
  role "ADMIN"; firstName/lastName/fullName unchanged, no leaked `name` property

One deliberate wire change: Product.categoryIds. A raw ObjectId serialized as
{"timestamp":…,"date":…}, an unusable shape; it is now the hex string. Verified by
grep across both frontends that neither reads categoryIds — they consume
categoryNames and send category ids back as the `genres` query parameter, already
as hex.

One deliberate behaviour change: an unrecognized `condition` filter is now a 400
instead of a string-equality match that could only ever return nothing. The
storefront only ever sends values from its own fixed list.

Two nulls that Money's scale makes visible and neither of which changes an outcome:
the MoMo gateway now receives BigDecimal("36.60") where it received ("36.6") —
MomoService scales to whole VND before signing — and a cart line's condition is
normalized to the enum's token rather than stored with the caller's casing.

240 tests green (147 before the sprint).

+++ backend/src/main/java/com/dominator/gearly/ai/CustomerServiceHandler.java
@@ -31,7 +31,7 @@ public class CustomerServiceHandler {
             // 2. Build the product context for the main bot
             String productContext = products.stream()
                 .map(p -> String.format("- %s (ID: %s): Price %s, Rating %s",
-                    p.getTitle(), p.getId(), p.getPrice(), p.getAverageRating()))
+                    p.getTitle(), p.getId(), p.getPrice().toDouble(), p.getAverageRating()))
                 .collect(Collectors.joining("\n"));
 
             String prompt = prompts.getProductRecommendation()
